+++ src/main/java/com/find/doongji/address/payload/request/AddressMapping.java
@@ -16,7 +16,7 @@ public class AddressMapping {
     private String umdNm;
     private String jibun;
     private String aptSeq;
-    private int danjiId;
+    private Long danjiId;
     private String bjdCode;
     private String aptNm;
